Refine caching behavior
All responses are now fetched from the network first, and use the SW
cache as a fallback only (stale resources were becoming an issue, and
testing indicates that serving from cache wasn't any faster, possibly
because the function that was handling fetches was waiting for the
network response to finish before returning the cached data - whoops).
The cache is exclusively for offline now.

Requests that need no special treatment or caching (ie, mp4 files) no
longer go through the service worker at all. I thought just responding
to the event with a normal fetch promise would work as a pass-through,
but apparently that causes it to try and buffer the entire response
before playing anything, and the correct solution is to just not
handle the event at all and the browser will do it for you normally.